requirements.txt
sentient-agent-framework==0.3.0


aiohttp==3.9.1


python-dotenv==1.0.0

